1129hb

+++ app/Http/Controllers/admins/ManagerController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
+use zgldh\QiniuStorage\QiniuStorage;
 use App\Http\Model\Manager;
 use session;
 use Hash;
@@ -70,18 +71,35 @@ class ManagerController extends Controller
         //文件上传
         if($request->hasFile('m_photo')){
 
-            //修改名字
-            $name = rand(1111,9999).time();
-            //获取后缀
-            $suffix = $request->file('m_photo')->getClientOriginalExtension();
-            //移动图片
-            $request->file('m_photo')->move('./Uploads',$name.'.'.$suffix);
+        //实例化disk
+        $disk = QiniuStorage::disk('qiniu');
+
+        //获取图片文件信息
+         $file = $request->file('m_photo');
+
+        //获取后缀
+        $suffix = $file->getClientOriginalExtension();
+
+        //拼装新的图片名称
+        $fileName = time().rand(100000,999999).'.'.$suffix;
+
+        //存进七牛
+        $bool=$disk->put('manager/'.$fileName,file_get_contents($file->getRealPath())); 
+
+
+/*
+        //修改名字
+        $name = rand(1111,9999).time();
+        //获取后缀
+        $suffix = $request->file('m_photo')->getClientOriginalExtension();
+        //移动图片
+        $request->file('m_photo')->move('./Uploads',$name.'.'.$suffix);*/
 
         }
 
         $res = $request->except('_token','m_photo','repass');
 
-        $res['m_photo'] = '/Uploads/'.$name.'.'.$suffix;
+        $res['m_photo'] = 'manager/'.$fileName;
 
         $res['m_password'] = Hash::make($request->input('m_password'));
 
@@ -144,19 +162,110 @@ class ManagerController extends Controller
      */
     public function destroy($id)
     {
-        $res = Manager::where('id',$id)->first();
 
-        $data = unlink('.'.$res->m_photo);
+        $mid = session('mid');
 
-        // var_dump($data);
+        if ($mid != $id) {
+            $m_photo = Manager::where('id',$id)->first()['m_photo'];
 
-        if ($data) {
-          $info =  Manager::where('id',$id)->delete();
-          if ($info) {
-              return redirect('/admin/manager')->with('msg','删除成功');
-          } else {
-            return back();
-          }
+            //实例七牛对象
+            $disk = QiniuStorage::disk('qiniu');
+            //删除七牛云图片
+            $data = $disk->delete($m_photo);
+
+            // var_dump($data);
+
+            if ($data) {
+              $info =  Manager::where('id',$id)->delete();
+              if ($info) {
+                  return redirect('/admin/manager')->with('msg','删除成功');
+              } else {
+                return back();
+              }
+            }
+        } else {
+            return redirect('/admin/manager')->with('msg','不能删除当前用户');
         }
+
+        
     }
+
+    //退出
+    public function logout(Request $request)
+    {
+        $request->session()->forget('mid');
+        return redirect('/admin/login');
+
+    }
+
+
+    //修改头像
+     public function editpic(Request $request)
+    {
+       
+        $mid = session('mid');
+        $m_name = Manager::where('id',$mid)->first()['m_name'];
+        return view('admins.manager.editpic',['m_name'=>$m_name]);
+
+    }
+
+    //执行修改头像
+     public function doedit(Request $request)
+    {
+
+        //文件上传
+                if($request->hasFile('m_photo')){
+
+                $mid = session('mid');
+                $m_photo = Manager::where('id',$mid)->first()['m_photo'];
+
+                //实例七牛对象
+                $disk = QiniuStorage::disk('qiniu');
+                // dd($m_photo);
+                //删除七牛云图片
+                $res = $disk->delete($m_photo);
+
+                 // dd($res);
+
+                if ($res) {
+
+                    
+
+                //获取图片文件信息
+                 $file = $request->file('m_photo');
+
+                //获取后缀
+                 $suffix = $file->getClientOriginalExtension();
+
+                //拼装新的图片名称
+                 $fileName = time().rand(100000,999999).'.'.$suffix;
+
+                //存进七牛
+                  $bool=$disk->put('manager/'.$fileName,file_get_contents($file->getRealPath()));
+
+
+                //存进数据库、
+                 $bool1 = Manager::where('id',$mid)->update(['m_photo'=>'manager/'.$fileName]);
+
+
+
+                if ($bool && $bool1) {
+                    return redirect('/admin/manager');
+                } else {
+                    return  back()->withInput();
+                }
+
+            } else {
+                return  back()->withInput();
+            }
+
+
+        } else {
+            return  back()->withInput();
+        }
+
+        
+
+    }
+
 }
